+++ addons/ezcha_network/lib/dto/trophy_obtained.gd
@@ -1,0 +1,5 @@
+extends EzchaTrophy
+class_name EzchaTrophyObtained
+
+## The timestamp for when the user obtained the trophy.
+var obtained_timestamp: String = ""